requirements.txt
pystray>=0.19.0
Pillow>=10.0.0
keyboard>=0.13.0
google-generativeai>=0.8.0
requests>=2.28.0
